Move TestOutputProvider to separate file
